Abstracting code to ContactListTableMobile
